kfs: tty:
fcntl64 impl
closexec flag is now per-task for fds, not in file
fixed bug with referencing files in dup
tty move left

+++ src/kernel/fs/inode.zig
@@ -93,7 +93,7 @@ pub const Inode = struct {
         return self.mode.canExecute(self.uid, self.gid);
     }
 
-    pub fn canAccess(self: *Inode, flags: u16) bool {
+    pub fn canAccess(self: *Inode, flags: u32) bool {
         if (kernel.task.current.uid == 0) {
             return true;
         }
